- modified directory structure by layer
