+++ src/main/java/edu/uml/diet/logic/FoodServiceException.java
@@ -1,9 +1,10 @@
-package edu.uml.diet;
+package edu.uml.diet.logic;
 
 /**
- * Used to signal an issue with PersistanceFoodService
+ * Created by Kurt Johnson on 3/8/2015.
  */
-public class PersistanceFoodServiceException extends Exception {
+public class FoodServiceException extends Exception {
+
     /**
      * Creates a new exception with a message passed as String to constructor
      * and cause for exception.
@@ -12,7 +13,7 @@ public class PersistanceFoodServiceException extends Exception {
      * @param cause cause for the exception, null value indicates cause is unknown
      *              or does not exist
      */
-    public PersistanceFoodServiceException(String message, Throwable cause) {
+    public FoodServiceException(String message, Throwable cause) {
         super(message, cause);
     }
 }
