+++ src/main/java/edu/uml/diet/logic/UserServiceException.java
@@ -1,10 +1,9 @@
-package edu.uml.diet;
+package edu.uml.diet.logic;
 
 /**
- * Created by Kurt Johnson on 3/8/2015.
+ * Created by Kurt Johnson on 3/3/2015.
  */
-public class FoodServiceException extends Exception {
-
+public class UserServiceException extends Exception {
     /**
      * Creates a new exception with a message passed as String to constructor
      * and cause for exception.
@@ -13,7 +12,7 @@ public class FoodServiceException extends Exception {
      * @param cause cause for the exception, null value indicates cause is unknown
      *              or does not exist
      */
-    public FoodServiceException(String message, Throwable cause) {
+    public UserServiceException(String message, Throwable cause) {
         super(message, cause);
     }
 }
