+++ src/main/java/edu/uml/diet/persistence/DatabaseConnector.java
@@ -1,4 +1,4 @@
-package edu.uml.diet;
+package edu.uml.diet.persistence;
 
 import java.sql.*;
