+++ src/main/java/edu/uml/diet/persistence/DuplicateFoodException.java
@@ -1,9 +1,10 @@
-package edu.uml.diet;
+package edu.uml.diet.persistence;
 
 /**
- * Created by Kurt Johnson on 3/3/2015.
+ * Used to signal that the BasicFood already exists in database
  */
-public class UserServiceException extends Exception {
+public class DuplicateFoodException extends Exception{
+
     /**
      * Creates a new exception with a message passed as String to constructor
      * and cause for exception.
@@ -12,7 +13,7 @@ public class UserServiceException extends Exception {
      * @param cause cause for the exception, null value indicates cause is unknown
      *              or does not exist
      */
-    public UserServiceException(String message, Throwable cause) {
+    public DuplicateFoodException(String message, Throwable cause) {
         super(message, cause);
     }
 }
