+++ src/main/java/edu/uml/diet/persistence/PersistanceFoodServiceException.java
@@ -1,10 +1,9 @@
-package edu.uml.diet;
+package edu.uml.diet.persistence;
 
 /**
- * Used to signal an issue with PersistanceUserServices
+ * Used to signal an issue with PersistanceFoodService
  */
-public class PersistanceUserServicesException extends Exception {
-
+public class PersistanceFoodServiceException extends Exception {
     /**
      * Creates a new exception with a message passed as String to constructor
      * and cause for exception.
@@ -13,7 +12,7 @@ public class PersistanceUserServicesException extends Exception {
      * @param cause cause for the exception, null value indicates cause is unknown
      *              or does not exist
      */
-    public PersistanceUserServicesException(String message, Throwable cause) {
+    public PersistanceFoodServiceException(String message, Throwable cause) {
         super(message, cause);
     }
 }
